VeilleAuto/requirements.txt
# requirements.txt
requests
beautifulsoup4
python-dotenv
google-generativeai>=0.5.0
